Separate LatexSubgraph
