+++ backend/src/airas/features/publication/generate_latex_subgraph/input_data.py
@@ -1,0 +1,1 @@
+generate_latex_subgraph_input_data: dict = {}
