+++ backend/src/airas/features/publication/generate_latex_subgraph/prompts/convert_to_latex_prompt.py
