--- backend/src/airas/features/publication/latex_subgraph/input_data.py
@@ -1,78 +1,0 @@
-from airas.types.github import GitHubRepositoryInfo
-from airas.types.paper import PaperContent
-
-latex_subgraph_input_data = {
-    "github_repository_info": GitHubRepositoryInfo(
-        github_owner="auto-res2",
-        repository_name="experiment_matsuzawa_chktex",
-        branch_name="develop",
-    ),
-    "paper_content": PaperContent(
-        title="Attention-Driven Reinforcement Learning for Dynamic Environments",
-        abstract="In this paper, we present a novel approach that integrates transformer-based multi-head attention with reinforcement learning to address the challenges of decision-making in dynamic and high-dimensional environments. Our method leverages the strengths of attention mechanisms to capture crucial temporal dependencies and uses policy gradient techniques for robust optimization. Traditional reinforcement learning methods often struggle with long-term dependency modeling and delayed reward signals, making it difficult to extract relevant historical information. By combining the attention mechanism, as popularized in [ashish_2017_attention], with policy gradient strategies reminiscent of [schulman_2017_ppo], our approach significantly improves the agent's ability to focus on informative past events, leading to a 15% increase in average cumulative reward compared to leading baselines. We validate our method on a set of standard benchmarks including Atari games and continuous control tasks, using evaluation metrics based on average cumulative rewards over 100 episodes. Detailed ablation studies and visual examinations of attention weights further support the efficacy of our framework. These empirical findings underline the potential of attention-driven reinforcement learning to enhance stability and performance in complex decision-making scenarios.",
-        introduction="Reinforcement learning has emerged as a powerful framework for training agents to perform complex sequential decision-making tasks in dynamic environments. Nevertheless, existing methods face significant challenges when it comes to capturing long-term dependencies, especially in settings characterized by high-dimensional inputs and delayed rewards. Our work addresses these challenges through the integration of transformer-based attention mechanisms into a reinforcement learning framework. The key innovation lies in the deployment of multi-head attention, as introduced in [ashish_2017_attention], to selectively encode temporal dependencies from past states, which are then utilized by a policy network optimized using policy gradient techniques inspired by [schulman_2017_ppo]. This approach enhances the model's capacity to focus on salient features of past observations, thereby overcoming shortcomings inherent in methods like Deep Q-Networks [mnih_2015_dqn] and traditional policy gradient algorithms that do not explicitly model long-range dependencies.\n\nThe significance of this research is multifold. First, by bridging attention mechanisms with reinforcement learning, we provide a solution to the credit assignment problem inherent in environments with delayed rewards. Second, our methodology improves the agent's performance through more precise representation learning, which leads to enhanced decision-making. Third, our extensive evaluation on diverse benchmark tasks demonstrates the practical effectiveness of our approach, with empirical results showing a consistent 15% improvement in cumulative rewards across various environments.\n\nOur contributions can be summarized as follows:\n- We propose an innovative reinforcement learning framework that synergistically combines multi-head attention with policy gradient optimization, thereby enabling more effective temporal representation learning.\n- We perform a comprehensive evaluation of our method on a variety of benchmark environments, including Atari games and continuous control tasks, and demonstrate its superiority over established baselines such as Proximal Policy Optimization [schulman_2017_ppo].\n- We conduct detailed ablation studies that underscore the importance of both the attention mechanism and the reinforcement learning components, validating that the removal of either results in significant performance degradation.\n- We provide visual interpretations of attention weights that offer critical insights into the decision-making process of the network, highlighting its focus on relevant historical states.\n\nThe remainder of the paper is structured as follows. In the next section, we review relevant literature, drawing comparisons with alternative approaches. We then introduce the necessary background and formalism that underpin our method. Following this, the technical details of our approach are elaborated, and the experimental setup is described in depth. Our results are subsequently presented and analyzed, and we conclude with a summary of our findings and a discussion of potential future research directions.",
-        related_work="Previous research in reinforcement learning has primarily focused on combining deep neural networks with traditional value-based or policy-gradient methods, with notable early contributions including the Deep Q-Network (DQN) [mnih_2015_dqn]. DQN demonstrated that deep learning could be used to directly learn control policies from raw sensory input, but it struggled with long-term temporal dependencies and delayed rewards. More recently, methods such as Proximal Policy Optimization (PPO) [schulman_2017_ppo] have been developed to improve training stability and sample efficiency by adopting surrogate objective functions. Unlike these approaches, our method explicitly models temporal dependencies via an integrated attention mechanism, which is central to the transformer architecture [ashish_2017_attention]. Transformers, although initially designed for sequence transduction problems in natural language processing, provide a robust mechanism for focusing on relevant parts of an input sequence. \n\nSeveral recent studies have explored extensions or hybridizations of standard reinforcement learning techniques to better capture temporal relationships, yet many such methods either rely on recurrent neural networks or lack the scalability offered by attention mechanisms. Approaches using recurrent units can suffer from vanishing gradients and limited capacity in long sequences, while our model leverages multi-head attention to maintain a broader contextual awareness. Furthermore, while there have been attempts to incorporate attention layers into reinforcement learning pipelines, these works generally consider them as auxiliary components rather than integral to the control policy. Our work differentiates itself by fully integrating attention into the policy formation process, thereby directly addressing the limitations observed in earlier models.\n\nAdditionally, while some literature has compared the capabilities of these different frameworks in isolated environments, our paper offers a comprehensive experimental evaluation across multiple standard benchmarks. By directly comparing our method against state-of-the-art baselines, such as PPO, we provide a clear and rigorous analysis of the strengths and limitations of the attention-driven approach. In summary, although there is a substantial body of related work addressing temporal dependencies and decision-making in reinforcement learning, the explicit combination of transformer-based attention with policy gradient methods, as implemented in our framework, presents a novel and promising direction for future research.",
-        background="The theoretical foundation of our work lies at the intersection of reinforcement learning and attention mechanisms, with a particular focus on the challenges associated with temporal credit assignment in dynamic environments. In reinforcement learning, the objective is to learn a policy π that maximizes the expected cumulative reward, where the decision-making process is influenced by both immediate and delayed rewards. Traditional methods such as DQN [mnih_2015_dqn] and PPO [schulman_2017_ppo] have achieved considerable success in this regard but often fall short when tasked with capturing long-range dependencies in sequential data. \n\nThe transformer model, introduced in [ashish_2017_attention], revolutionized sequence modeling by utilizing multi-head attention mechanisms to compute self-attention across input sequences. This process involves the computation of similarity scores between different states, normalizing these scores, and then generating a weighted representation that emphasizes the most relevant features. In the context of reinforcement learning, this allows the model to dynamically allocate attention to parts of the input sequence that are critical for decision-making, even when those inputs are separated by long time intervals.\n\nOur problem setting can be formally stated as follows. Given a state space S and an action space A, the goal is to learn a policy function π: S → A that maps environmental states to actions that maximize the expected return. The integration of attention into this framework introduces an additional representation layer, whereby the input state sequence is transformed into an embedding that reflects temporal dependencies. Key assumptions include the stationarity of the environment over limited time frames and the sufficiency of random seed initializations for providing diverse starting conditions during training. \n\nBy fusing these ideas, our framework not only addresses the long-standing challenge of delayed rewards but also enables a more nuanced understanding of the sequential patterns present in dynamic environments. This deeper insight into the temporal structure of the environment paves the way for more efficient and effective policy optimization.",
-        method="Our method is built upon the integration of transformer-based multi-head attention within a reinforcement learning framework to manage temporal dependencies and challenges associated with delayed rewards. The approach is comprised of two main components: a transformer encoder that employs multi-head attention to process sequences of previous states, and a policy network that uses the resulting embeddings to select actions. \n\nIn the first step, the agent collects a sequence of states from the environment. This sequence is then input into a transformer encoder module where multiple attention heads compute similarity scores among the states. These scores are normalized using a softmax function, effectively assigning weights that highlight the contributions of certain past states over others. This process, inspired by the architecture in [ashish_2017_attention], is critical in providing a robust representation of the temporal context. \n\nIn the next stage, the output embeddings from the attention module serve as refined input to the policy network. The policy network utilizes these informative embeddings to calculate a distribution over the available actions, selecting the most appropriate action based on the contextual information. To optimize the policy network, we adopt a policy gradient strategy akin to that described in [schulman_2017_ppo]. The optimization procedure involves sampling trajectories from the environment, computing cumulative rewards, and adjusting the network parameters via stochastic gradient ascent using the Adam optimizer with a fixed learning rate of 3e-4 and a batch size of 256. \n\nThe overall training procedure is conducted over 1 million timesteps with periodic evaluations every 10,000 steps. A typical training cycle consists of resetting the environment, allowing the agent to interact with it while recording state transitions, and updating the policy based on the observed rewards. We provide a pseudocode outline to illustrate the process:\n\n1. Initialize the environment and model parameters.\n2. For each timestep up to 1 million, perform:\n   a. Reset the environment to obtain the initial state.\n   b. Process the current state sequence through the transformer encoder to obtain softened representation via multi-head attention.\n   c. Use the policy network to select an action based on the attention-enhanced embedding.\n   d. Execute the action and obtain the subsequent state and reward.\n   e. Update the policy network with the observed transition using the policy gradient method.\n   f. Every 10,000 steps, evaluate the model's performance over a fixed number of episodes.\n\nA critical aspect of our method is the ablation study wherein we remove either the attention module or the reinforcement learning component to gauge their individual contributions. Our experiments clearly indicate that the removal of either component results in a significant decline in performance. This underscores the necessity of their combined operation for achieving the observed improvements in cumulative reward. The methodological choices, including specific implementation details such as the choice of optimizer, learning rate, and batch size, have been carefully validated through experimental runs.",
-        experimental_setup="Our experimental design is tailored to rigorously evaluate the performance of the proposed attention-based reinforcement learning framework. The experiments are executed on five distinct environments, including popular Atari games and continuous control tasks, each initialized with three separate random seeds. The diverse set of environments ensures the robustness of our findings across different types of dynamic decision-making scenarios. \n\nFor each environment, the training is conducted for 1 million timesteps with performance evaluations performed every 10,000 steps. The agent processes the current state along with a historical sequence of states through a transformer encoder that applies multi-head attention to extract temporal patterns. The resultant embedding is fed into a policy network that selects actions according to a learned probability distribution. The network parameters are updated using the Adam optimizer, with a fixed learning rate of 3e-4 and a batch size set to 256. \n\nKey evaluation metrics include the average cumulative reward measured over 100 episodes. This metric provides a comprehensive indication of the agent's performance and its ability to exploit the learned policy. Moreover, we perform ablation studies by altering the network structure – specifically by removing either the attention component or using a standard reinforcement learning model – to subjectively assess the contribution of the attention mechanism. \n\nThe pseudocode for the training procedure is as follows:\n\n• Define a training function that accepts the environment, model, and number of timesteps.\n• Initialize the optimizer (Adam) with a learning rate of 3e-4.\n• For each timestep from 1 to 1,000,000:\n    o Reset the environment to obtain the initial state.\n    o While the episode is not terminated, select an action using the current model via the attention-enhanced policy network.\n    o After executing the action, record the next state, reward, and termination signal.\n    o Update the model parameters using the observed transition data.\n    o Increment the timestep counter and update cumulative rewards.\n    o Every 10,000 timesteps, evaluate the model to get a snapshot of its performance.\n\nBaseline methods such as Proximal Policy Optimization [schulman_2017_ppo] are utilized for performance comparison. Additionally, experiments are run on tasks like Breakout, SpaceInvaders, and CartPole to capture a range of environmental dynamics. The hyperparameter settings described above are maintained consistently across all experiments to ensure fair comparisons. All datasets and tasks are sourced from standard benchmarks used in the reinforcement learning community, and evaluation protocols adhere to established practices, thereby ensuring reproducibility and transparency in our experimental design.",
-        results="Our experimental analysis shows that integrating transformer-based attention within a reinforcement learning framework yields significant performance gains. Across all tested environments, our method achieves an average cumulative reward that is approximately 15% higher than the best-performing baseline, Proximal Policy Optimization [schulman_2017_ppo]. Specifically, the performance metrics recorded during our tests indicate that the agent obtained an average reward of 520 ± 25 points on Breakout, 1840 ± 67 points on SpaceInvaders, and 95 ± 8 points on CartPole. These improvements confirm the efficacy of the attention module in capturing vital temporal dependencies that standard reinforcement learning methods often overlook.\n\nDetailed analysis reveals that the training dynamics are notably stable, with convergence curves consistently trending upwards over the training period. Moreover, visualizations of the attention weights indicate that the model effectively focuses on historical states that are most influential for current decision-making, thereby validating our approach. Our ablation studies further underline this point; when the attention mechanism is removed, the performance degrades significantly, confirming the synergistic benefit of the combined architecture.\n\nFigure 1: Training Curves – This figure (filename: training_curves.pdf) illustrates the progression and convergence of the cumulative rewards over time, demonstrating stable performance improvements at regular evaluation intervals.\n\nFigure 2: Attention Visualization – This figure (filename: attention_visualization.pdf) provides a visual representation of the temporal attention weights learned by the model, clearly highlighting the focus on relevant past states during decision-making.\n\nFigure 3: Performance Comparison – This figure (filename: performance_comparison.pdf) offers a comparative analysis between our method and baseline approaches, visually emphasizing the approximately 15% increase in performance across the tested environments.\n\nThese results have been consistently replicated across different random seeds, reinforcing the generalizability of our approach. Limitations of the method include the computational overhead introduced by the attention mechanism, although this has not hindered its scalability with respect to larger environments. Overall, our experiments suggest that the integrated model not only offers quantitative improvements but also provides enhanced qualitative insights into the decision-making process through the interpretation of attention weights.",
-        conclusion="This paper has introduced an innovative reinforcement learning framework that integrates multi-head attention with policy gradient methods to address the challenges of dynamic, high-dimensional environments characterized by long-term dependencies and delayed rewards. By leveraging the transformer architecture as outlined in [ashish_2017_attention] alongside techniques from [schulman_2017_ppo], our method enhances the ability of the agent to focus on relevant historical states, thereby achieving a consistent 15% improvement in average cumulative rewards compared to established baselines. Extensive experiments conducted on environments such as Breakout, SpaceInvaders, and CartPole have validated the proposed approach, with ablation studies further confirming the essential role of both the attention mechanism and the reinforcement learning components. \n\nIn addition to demonstrating quantitative improvements, our work provides qualitative insights through visualizations of temporal attention weights, offering a deeper understanding of the decision-making process. While the integration of attention adds computational overhead, the resulting gains in performance and stability justify this trade-off. \n\nFuture work will focus on scaling the approach to even more complex environments and exploring alternative attention architectures that may further reduce computational costs. Overall, our findings underscore the potential of attention-driven reinforcement learning as a powerful tool for tackling challenging dynamic decision-making problems, paving the way for further research in this promising direction.",
-    ),
-    "references_bib": """% ===========================================
-% REQUIRED CITATIONS
-% These papers must be cited in the manuscript
-% ===========================================
-
-@article{john_2023_deep,
- abstract = {This paper presents a comprehensive survey of deep learning techniques applied to natural language processing tasks, including sentiment analysis, machine translation, and text classification.},
- arxiv_url = {https://arxiv.org/abs/2301.12345},
- author = {Smith, John and Johnson, Alice},
- doi = {10.1038/s42256-023-00567-8},
- github_url = {https://github.com/nlp-survey/deep-learning-nlp},
- journal = {Nature Machine Intelligence},
- number = {3},
- pages = {123-145},
- title = {Deep Learning for Natural Language Processing: A Comprehensive Survey},
- volume = {5},
- year = {2023}
-}
-
-@article{michael_2024_transformer,
- abstract = {We explore the transformer architecture and its applications across various domains including computer vision and natural language processing.},
- arxiv_url = {https://arxiv.org/abs/2401.56789},
- author = {Brown, Michael and Davis, Sarah},
- doi = {10.5555/3648699.3648700},
- journal = {Journal of Machine Learning Research},
- pages = {1-28},
- title = {Transformer Networks: Architecture and Applications},
- volume = {25},
- year = {2024}
-}
-
-% ===========================================
-% REFERENCE CANDIDATES
-% Additional reference papers for context
-% ===========================================
-
-@article{ashish_2017_attention,
- abstract = {The dominant sequence transduction models are based on complex recurrent or convolutional neural networks that include an encoder and a decoder. The best performing models also connect the encoder and decoder through an attention mechanism.},
- arxiv_url = {https://arxiv.org/abs/1706.03762},
- author = {Vaswani, Ashish and Shazeer, Noam and Parmar, Niki},
- doi = {10.48550/arXiv.1706.03762},
- journal = {Advances in Neural Information Processing Systems},
- title = {Attention Is All You Need},
- volume = {30},
- year = {2017}
-}
-
-@article{b_2020_gpt,
- abstract = {Recent work has demonstrated substantial gains on many NLP tasks and benchmarks by pre-training on a large corpus of text followed by fine-tuning on a specific task.},
- arxiv_url = {https://arxiv.org/abs/2005.14165},
- author = {Brown, Tom B. and Mann, Benjamin and Ryder, Nick},
- journal = {Advances in Neural Information Processing Systems},
- pages = {1877-1901},
- title = {GPT-3: Language Models are Few-Shot Learners},
- volume = {33},
- year = {2020}
-}""",
-}
