+++ backend/src/airas/features/publication/publish_latex_subgraph/input_data.py
@@ -1,0 +1,1 @@
+publish_latex_subgraph_input_data: dict = {}
