+++ backend/src/airas/features/publication/publish_latex_subgraph/nodes/embed_in_latex_template.py
