+++ backend/src/airas/types/latex.py
@@ -1,3 +1,12 @@
 from typing import Literal
 
+from airas.types.github import GitHubRepositoryInfo
+
 LATEX_TEMPLATE_NAME = Literal["iclr2024", "agents4science_2025"]
+
+# Official LaTeX template repository
+LATEX_TEMPLATE_REPOSITORY_INFO = GitHubRepositoryInfo(
+    github_owner="airas-org",
+    repository_name="airas-template",
+    branch_name="main",
+)
